\name{DJIA}
\alias{DJIA}
\docType{data}
\title{Dow Jones data}
\description{
A dataset for testing. It contains the weekly open price of the Dow Jones Industrial Average (DJIA) stock market indexfrom the first week in September, 2000 till the last week in August, 2002.}

\format{This data frame contains only one column: open
}

\examples{
data(DJIA)
}
\keyword{datasets}